# docs/reference/backend-api/openapi.generated.yaml
swagger: '2.0'
info:
  title: Pipeline Auth Live Router API Reference
  description: >-
    Generated from live router definitions and merged with the published Swagger artifact when matching operations
    exist.
  version: 1.503.0
  contact:
    email: casbin@googlegroups.com
basePath: /
schemes:
  - https
  - http
paths:
...
  /api/get-user:
    get:
      tags:
        - User API
      description: get user
      operationId: ApiController.GetUser
      parameters:
        - in: query
          name: id
          description: The id ( owner/name ) of the user
          type: string
        - in: query
          name: owner
          description: The owner of the user
          type: string
        - in: query
          name: email
          description: The email of the user
          type: string
        - in: query
          name: phone
          description: The phone of the user
          type: string
        - in: query
          name: userId
          description: The userId of the user
          type: string
      responses:
        '200':
          description: The Response object
          schema:
            $ref: '#/definitions/object.User'
    x-live-router-methods:
      - GET
    x-live-controller-actions:
      - ApiController.GetUser
    x-live-domains:
      - user
...
definitions:
...
  object.Address:
    title: Address
    type: object
    properties:
      city:
        type: string
      line1:
        type: string
      line2:
        type: string
      region:
        type: string
      state:
        type: string
      tag:
        type: string
      zipCode:
        type: string
...
  object.FaceId:
    title: FaceId
    type: object
    properties:
      ImageUrl:
        type: string
      faceIdData:
        type: array
        items:
          type: number
          format: double
      name:
        type: string
...
  object.ManagedAccount:
    title: ManagedAccount
    type: object
    properties:
      application:
        type: string
      password:
        type: string
      signinUrl:
        type: string
      username:
        type: string
  object.MfaAccount:
    title: MfaAccount
    type: object
    properties:
      accountName:
        type: string
      issuer:
        type: string
      origin:
        type: string
      secretKey:
        type: string
  object.MfaItem:
    title: MfaItem
    type: object
    properties:
      name:
        type: string
      rule:
        type: string
  object.MfaProps:
    title: MfaProps
    type: object
    properties:
      countryCode:
        type: string
      enabled:
        type: boolean
      isPreferred:
        type: boolean
      mfaRememberInHours:
        type: integer
        format: int64
      mfaType:
        type: string
      recoveryCodes:
        type: array
        items:
          type: string
      secret:
        type: string
      url:
        type: string
...
  object.Permission:
    title: Permission
    type: object
    properties:
      actions:
        type: array
        items:
          type: string
      adapter:
        type: string
      approveTime:
        type: string
      approver:
        type: string
      createdTime:
        type: string
      description:
        type: string
      displayName:
        type: string
      domains:
        type: array
        items:
          type: string
      effect:
        type: string
      groups:
        type: array
        items:
          type: string
      isEnabled:
        type: boolean
      model:
        type: string
      name:
        type: string
      owner:
        type: string
      resourceType:
        type: string
      resources:
        type: array
        items:
          type: string
      roles:
        type: array
        items:
          type: string
      state:
        type: string
      submitter:
        type: string
      users:
        type: array
        items:
          type: string
...
  object.ProductInfo:
    title: ProductInfo
    type: object
    properties:
      currency:
        type: string
      detail:
        type: string
      displayName:
        type: string
      image:
        type: string
      isRecharge:
        type: boolean
      name:
        type: string
      owner:
        type: string
      planName:
        type: string
      price:
        type: number
        format: double
      pricingName:
        type: string
      quantity:
        type: integer
        format: int64
...
  object.Role:
    title: Role
    type: object
    properties:
      createdTime:
        type: string
      description:
        type: string
      displayName:
        type: string
      domains:
        type: array
        items:
          type: string
      groups:
        type: array
        items:
          type: string
      isEnabled:
        type: boolean
      name:
        type: string
      owner:
        type: string
      roles:
        type: array
        items:
          type: string
      users:
        type: array
        items:
          type: string
...
  object.User:
    title: User
    type: object
    properties:
      accessKey:
        type: string
      accessSecret:
        type: string
      accessToken:
        type: string
      address:
        type: array
        items:
          type: string
      addresses:
        type: array
        items:
          $ref: '#/definitions/object.Address'
      adfs:
        type: string
      affiliation:
        type: string
      alipay:
        type: string
      amazon:
        type: string
      apple:
        type: string
      auth0:
        type: string
      avatar:
        type: string
      avatarType:
        type: string
      azuread:
        type: string
      azureadb2c:
        type: string
      baidu:
        type: string
      balance:
        type: number
        format: double
      balanceCredit:
        type: number
        format: double
      balanceCurrency:
        type: string
      battlenet:
        type: string
      bilibili:
        type: string
      bio:
        type: string
      birthday:
        type: string
      bitbucket:
        type: string
      box:
        type: string
      cart:
        type: array
        items:
          $ref: '#/definitions/object.ProductInfo'
      casdoor:
        type: string
      cloudfoundry:
        type: string
      countryCode:
        type: string
      createdIp:
        type: string
      createdTime:
        type: string
      currency:
        type: string
      custom:
        type: string
      custom2:
        type: string
      custom3:
        type: string
      custom4:
        type: string
      custom5:
        type: string
      custom6:
        type: string
      custom7:
        type: string
      custom8:
        type: string
      custom9:
        type: string
      custom10:
        type: string
      dailymotion:
        type: string
      deezer:
        type: string
      deletedTime:
        type: string
      digitalocean:
        type: string
      dingtalk:
        type: string
      discord:
        type: string
      displayName:
        type: string
      douyin:
        type: string
      dropbox:
        type: string
      education:
        type: string
      email:
        type: string
      emailVerified:
        type: boolean
      eveonline:
        type: string
      externalId:
        type: string
      faceIds:
        type: array
        items:
          $ref: '#/definitions/object.FaceId'
      facebook:
        type: string
      firstName:
        type: string
      fitbit:
        type: string
      gender:
        type: string
      gitea:
        type: string
      gitee:
        type: string
      github:
        type: string
      gitlab:
        type: string
      google:
        type: string
      groups:
        type: array
        items:
          type: string
      hash:
        type: string
      heroku:
        type: string
      homepage:
        type: string
      id:
        type: string
      idCard:
        type: string
      idCardType:
        type: string
      influxcloud:
        type: string
      infoflow:
        type: string
      instagram:
        type: string
      intercom:
        type: string
      invitation:
        type: string
      invitationCode:
        type: string
      ipWhitelist:
        type: string
      isAdmin:
        type: boolean
      isDefaultAvatar:
        type: boolean
      isDeleted:
        type: boolean
      isForbidden:
        type: boolean
      isOnline:
        type: boolean
      isVerified:
        type: boolean
      kakao:
        type: string
      karma:
        type: integer
        format: int64
      kwai:
        type: string
      language:
        type: string
      lark:
        type: string
      lastChangePasswordTime:
        type: string
      lastName:
        type: string
      lastSigninIp:
        type: string
      lastSigninTime:
        type: string
      lastSigninWrongTime:
        type: string
      lastfm:
        type: string
      ldap:
        type: string
      line:
        type: string
      linkedin:
        type: string
      location:
        type: string
      mailru:
        type: string
      managedAccounts:
        type: array
        items:
          $ref: '#/definitions/object.ManagedAccount'
      meetup:
        type: string
      metamask:
        type: string
      mfaAccounts:
        type: array
        items:
          $ref: '#/definitions/object.MfaAccount'
      mfaEmailEnabled:
        type: boolean
      mfaItems:
        type: array
        items:
          $ref: '#/definitions/object.MfaItem'
      mfaPhoneEnabled:
        type: boolean
      mfaPushEnabled:
        type: boolean
      mfaPushProvider:
        type: string
      mfaPushReceiver:
        type: string
      mfaRadiusEnabled:
        type: boolean
      mfaRadiusProvider:
        type: string
      mfaRadiusUsername:
        type: string
      mfaRememberDeadline:
        type: string
      microsoftonline:
        type: string
      multiFactorAuths:
        type: array
        items:
          $ref: '#/definitions/object.MfaProps'
      name:
        type: string
      naver:
        type: string
      needUpdatePassword:
        type: boolean
      nextcloud:
        type: string
      okta:
        type: string
      onedrive:
        type: string
      originalRefreshToken:
        type: string
      originalToken:
        type: string
      oura:
        type: string
      owner:
        type: string
      password:
        type: string
      passwordSalt:
        type: string
      passwordType:
        type: string
      patreon:
        type: string
      paypal:
        type: string
      permanentAvatar:
        type: string
      permissions:
        type: array
        items:
          $ref: '#/definitions/object.Permission'
      phone:
        type: string
      preHash:
        type: string
      preferredMfaType:
        type: string
      properties:
        additionalProperties:
          type: string
      qq:
        type: string
      ranking:
        type: integer
        format: int64
      realName:
        type: string
      recoveryCodes:
        type: array
        items:
          type: string
      region:
        type: string
      registerSource:
        type: string
      registerType:
        type: string
      roles:
        type: array
        items:
          $ref: '#/definitions/object.Role'
      salesforce:
        type: string
      score:
        type: integer
        format: int64
      shopify:
        type: string
      signinWrongTimes:
        type: integer
        format: int64
      signupApplication:
        type: string
      slack:
        type: string
      soundcloud:
        type: string
      spotify:
        type: string
      steam:
        type: string
      strava:
        type: string
      stripe:
        type: string
      tag:
        type: string
      tiktok:
        type: string
      title:
        type: string
      totpSecret:
        type: string
      tumblr:
        type: string
      twitch:
        type: string
      twitter:
        type: string
      type:
        type: string
      typetalk:
        type: string
      uber:
        type: string
      updatedTime:
        type: string
      vk:
        type: string
      web3onboard:
        type: string
      webauthnCredentials:
        type: array
        items:
          $ref: '#/definitions/webauthn.Credential'
      wechat:
        type: string
      wecom:
        type: string
      weibo:
        type: string
      wepay:
        type: string
      xero:
        type: string
      yahoo:
        type: string
      yammer:
        type: string
      yandex:
        type: string
      zoom:
        type: string
...
  webauthn.Credential:
    title: Credential
    type: object
...
securityDefinitions:
  AccessToken:
    type: apiKey
    name: Authorization
    in: header
...